+++ tests/testthat/test-link-href.R
@@ -54,7 +54,7 @@ test_that("can link to functions in base packages", {
   scoped_package_context("test")
   scoped_file_context() # package registry maintained on per-file basis
 
-  expect_equal(href_expr_(library()), href_topic_remote("library", "base"))
+  expect_equal(href_expr_(abbreviate()), href_topic_remote("abbreviate", "base"))
   expect_equal(href_expr_(median()), href_topic_remote("median", "stats"))
 })
 
@@ -106,16 +106,25 @@ test_that("can link ? calls", {
   expect_equal(href_expr_(package?foo), "foo-package.html")
 })
 
-
 test_that("can link help calls", {
   scoped_package_context("test", c(foo = "foo", "foo-package" = "foo-package"))
   scoped_file_context("bar")
 
   expect_equal(href_expr_(help("foo")), "foo.html")
   expect_equal(href_expr_(help("foo", "test")), "foo.html")
+  expect_equal(href_expr_(help(package = "MASS")), "https://rdrr.io/pkg/MASS/man")
   expect_equal(href_expr_(help()), NA_character_)
 })
 
+
+# library and friends -----------------------------------------------------
+
+test_that("library() linked to package reference", {
+  scoped_package_context("test", c(foo = "bar"))
+
+  expect_equal(href_expr_(library(pkgdown)), "https://pkgdown.r-lib.org/reference")
+  expect_equal(href_expr_(library(MASS)), "https://rdrr.io/pkg/MASS/man")
+})
 
 # vignette ----------------------------------------------------------------